requirements.txt
amqp==5.1.1
APScheduler==3.6.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cachetools==4.2.2
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.1.1
django-admin-rangefilter==0.9.0
django-celery-beat==2.4.0
django-environ==0.9.0
django-import-export==2.9.0
django-phonenumber-field==7.0.0
django-timezone-field==5.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==1.2.0
idna==3.4
ipython==8.7.0
jedi==0.18.2
kombu==5.2.4
MarkupPy==1.14
matplotlib-inline==0.1.6
odfpy==1.4.1
openpyxl==3.0.10
packaging==22.0
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.12.54
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyTelegramBotAPI==4.7.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-telegram-bot==13.15
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tablib==3.3.0
tornado==6.1
traitlets==5.8.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
